Add main function for testing speed
